+++ CHANGELOG.md
@@ -4,12 +4,20 @@ All Notable changes to `html-tokenizer` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## [0.2.1] - 2016-05-08
+## 2016-XX-XX
+
+### Added
+- Contents of "iframe" element ignored.
+
+### Todo
+- Collapse whitespace around text correctly.
+
+## [0.2.1] - 2016-05-07
 
 ### Fixed
 - Bad regex for PHP token.
 
-## [0.2.0] - 2016-05-08
+## [0.2] - 2016-05-07
 
 ### Added
 - Added PHP token.
@@ -18,7 +26,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Changed
 - Use mb_* string functions.  Polyfill added.
 
-## [0.1.0] - 2016-04-17
+## [0.1] - 2016-04-17
 
 ### Added
 - Initial release